Add multiprocessing to generator
